Split product-v2-form into per-section components
The form had grown past 1500 lines as a single component. Split into:
- product-v2-form.tsx — orchestrator (state, validate, submit, render)
- product-v2-form-state.ts — FormState, initialState, derivation helpers
- form-primitives.tsx — shared FormSection / Field / InfoCallout
- sections/* — one file per form section (identity, audience, where,
  when, groups, billing, registration, visibility)

Each section receives {state, setState, config?, ...} and calls its own
React-Query hooks (deduped at the query layer). Multi-line derivations
(effectiveBillingMode, locationPickerMode, etc.) live in the state
module so parent validate/submit and consuming sections share one
source of truth.

No behavior changes.
